Cleaning up JunHong UI Prototype Branch
Prototypes before 3 and Model training results before V5 are removed.

Added UI_prototype
