--- HandFlux/Prototype 1/ReadMe
@@ -1,2 +1,0 @@
-Link to my google drive folder contains the save model: gesture_recognition_model.h5
-https://drive.google.com/drive/folders/1pXE6p9h5AZwZGbxLDe7Nux_XXduGtBfS
