--- HandFlux/README.md
@@ -1,2 +1,0 @@
-Folder to the dataset used to train:
-https://drive.google.com/file/d/1QpIHZwzD3l0nXOb2Qc-_gdVwn6SGYWBn/view?usp=drive_link
