+++ UI_Prototype/faq_text.txt
@@ -1,0 +1,29 @@
+Q - How do I install this app?
+A - Simple, just extract the .exe from the zip folder, or run the .exe after using the installer.
+
+Q - Lorem Ipsum
+A - Dolor sit appmet
+
+Q - Consectetur adipiscing elit
+A - Sed do eiusmod
+
+Q - Tempor incididunt ut labore 
+A - Et dolore magna aliqua
+
+Q - Ut enim ad minim veniam
+A - Quis nostrud exercitation
+
+Q - Ullamco laboris nisi ut aliquip
+A - Ex ea commodo consequat
+
+Q - Duis aute irure dolor
+A - In reprehenderit in voluptate
+
+Q - Velit esse cillum dolore
+A - Eu fugiat nulla pariatur
+
+Q - 
+A -
+
+Q - 
+A -
