--- simple prototype/Kivy Sample.kv
@@ -1,3 +1,0 @@
-Label:
-    text: "Test"
-    font_size: 128
